fix: join ncor page create issue on repo
